v3.0 추가 (PyTorch 저용)
