+++ main.py
@@ -6,22 +6,22 @@ import json
 
 from quantylab.rltrader import settings
 from quantylab.rltrader import utils
-import data_manager
+from quantylab.rltrader import data_manager
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('--stock_code', nargs='+')
-    parser.add_argument('--ver', choices=['v1', 'v2', 'v3', 'v4'], default='v3')
+    parser.add_argument('--ver', choices=['v1', 'v2', 'v3', 'v4'], default='v2')
     parser.add_argument('--rl_method', choices=['dqn', 'pg', 'ac', 'a2c', 'a3c', 'monkey'])
-    parser.add_argument('--net', choices=['dnn', 'lstm', 'cnn', 'monkey'], default='dnn')
+    parser.add_argument('--net', choices=['dnn', 'lstm', 'cnn', 'q3', 'monkey'], default='dnn')
     parser.add_argument('--num_steps', type=int, default=1)
     parser.add_argument('--lr', type=float, default=0.001)
     parser.add_argument('--discount_factor', type=float, default=0.9)
     parser.add_argument('--start_epsilon', type=float, default=0)
     parser.add_argument('--balance', type=int, default=10000000)
     parser.add_argument('--num_epoches', type=int, default=100)
-    parser.add_argument('--backend', choices=['tensorflow', 'plaidml'], default='tensorflow')
+    parser.add_argument('--backend', choices=['pytorch', 'tensorflow', 'plaidml'], default='pytorch')
     parser.add_argument('--output_name', default=utils.get_time_str())
     parser.add_argument('--value_network_name')
     parser.add_argument('--policy_network_name')
@@ -31,7 +31,8 @@ if __name__ == '__main__':
     parser.add_argument('--end_date', default='20201231')
     args = parser.parse_args()
 
-    # Keras Backend 설정
+    # Backend 설정
+    os.environ['RLTRADER_BACKEND'] = args.backend
     if args.backend == 'tensorflow':
         os.environ['KERAS_BACKEND'] = 'tensorflow'
     elif args.backend == 'plaidml':
@@ -48,29 +49,31 @@ if __name__ == '__main__':
         f.write(json.dumps(vars(args)))
     
     # 로그 기록 설정
-    file_handler = logging.FileHandler(filename=os.path.join(
-        output_path, "{}.log".format(args.output_name)), encoding='utf-8')
+    log_path = os.path.join(output_path, '{}.log'.format(args.output_name))
+    if os.path.exists(log_path):
+        os.remove(log_path)
+    file_handler = logging.FileHandler(filename=log_path, encoding='utf-8')
+    file_handler.setLevel(logging.INFO)
     stream_handler = logging.StreamHandler(sys.stdout)
-    file_handler.setLevel(logging.DEBUG)
-    stream_handler.setLevel(logging.INFO)
-    logging.basicConfig(format="%(message)s",
-        handlers=[file_handler, stream_handler], level=logging.DEBUG)
+    stream_handler.setLevel(logging.WARNING)
+    logging.basicConfig(format="%(message)s", handlers=[file_handler, stream_handler], level=logging.INFO)
     
     # 로그, Keras Backend 설정을 먼저하고 RLTrader 모듈들을 이후에 임포트해야 함
     from quantylab.rltrader.learners import ReinforcementLearner, DQNLearner, \
         PolicyGradientLearner, ActorCriticLearner, A2CLearner, A3CLearner
 
     # 모델 경로 준비
+    # 모델 포멧은 TensorFlow는 h5, PyTorch는 pickle
     value_network_path = ''
     policy_network_path = ''
     if args.value_network_name is not None:
-        value_network_path = os.path.join(settings.BASE_DIR, 'models/{}.h5'.format(args.value_network_name))
+        value_network_path = os.path.join(settings.BASE_DIR, 'models/{}.mdl'.format(args.value_network_name))
     else:
-        value_network_path = os.path.join(output_path, '{}_{}_{}_value.h5'.format(args.output_name, args.rl_method, args.net))
+        value_network_path = os.path.join(output_path, '{}_{}_{}_value.mdl'.format(args.output_name, args.rl_method, args.net))
     if args.policy_network_name is not None:
-        policy_network_path = os.path.join(settings.BASE_DIR, 'models/{}.h5'.format(args.policy_network_name))
+        policy_network_path = os.path.join(settings.BASE_DIR, 'models/{}.mdl'.format(args.policy_network_name))
     else:
-        policy_network_path = os.path.join(output_path, '{}_{}_{}_policy.h5'.format(args.output_name, args.rl_method, args.net))
+        policy_network_path = os.path.join(output_path, '{}_{}_{}_policy.mdl'.format(args.output_name, args.rl_method, args.net))
 
     common_params = {}
     list_stock_code = []
