+++ quantylab/rltrader/agent.py
@@ -4,7 +4,7 @@ from quantylab.rltrader import utils
 
 class Agent:
     # 에이전트 상태가 구성하는 값 개수
-    # 주식 보유 비율, 포트폴리오 가치 비율, 평균 매수 단가 대비 등락률
+    # 주식 보유 비율, 현재 손익, 평균 매수 단가 대비 등락률
     STATE_DIM = 3
 
     # 매매 수수료 및 세금
@@ -23,66 +23,53 @@ class Agent:
     NUM_ACTIONS = len(ACTIONS)  # 인공 신경망에서 고려할 출력값의 개수
 
     def __init__(self, \
-        environment, balance, min_trading_unit=1, max_trading_unit=2):
-        # Environment 객체
+        environment, initial_balance, min_trading_unit=1, max_trading_unit=2):
         # 현재 주식 가격을 가져오기 위해 환경 참조
         self.environment = environment
+        self.initial_balance = initial_balance  # 초기 자본금
 
         # 최소 매매 단위, 최대 매매 단위, 지연보상 임계치
         self.min_trading_unit = min_trading_unit  # 최소 단일 거래 단위
         self.max_trading_unit = max_trading_unit  # 최대 단일 거래 단위
 
         # Agent 클래스의 속성
-        self.initial_balance = balance  # 초기 자본금
-        self.balance = balance  # 현재 현금 잔고
+        self.balance = initial_balance  # 현재 현금 잔고
         self.num_stocks = 0  # 보유 주식 수
         # 포트폴리오 가치: balance + num_stocks * {현재 주식 가격}
         self.portfolio_value = 0
-        self.base_portfolio_value = 0  # 직전 학습 시점의 PV
         self.num_buy = 0  # 매수 횟수
         self.num_sell = 0  # 매도 횟수
         self.num_hold = 0  # 홀딩 횟수
         self.immediate_reward = 0  # 즉시 보상
-        self.profitloss = 0  # 현재 손익
-        self.base_profitloss = 0  # 직전 지연 보상 이후 손익
-        self.exploration_base = 0  # 탐험 행동 결정 기준
 
         # Agent 클래스의 상태
         self.ratio_hold = 0  # 주식 보유 비율
-        self.ratio_portfolio_value = 0  # 포트폴리오 가치 비율
+        self.profitloss = 0  # 현재 손익
         self.avg_buy_price = 0  # 주당 매수 단가
 
     def reset(self):
         self.balance = self.initial_balance
         self.num_stocks = 0
         self.portfolio_value = self.initial_balance
-        self.base_portfolio_value = self.initial_balance
         self.num_buy = 0
         self.num_sell = 0
         self.num_hold = 0
         self.immediate_reward = 0
         self.ratio_hold = 0
-        self.ratio_portfolio_value = 0
-
-    def reset_exploration(self, alpha=None):
-        if alpha is None:
-            alpha = 0
-        self.exploration_base = 0.5 + alpha
+        self.profitloss = 0
+        self.avg_buy_price = 0
 
     def set_balance(self, balance):
         self.initial_balance = balance
 
     def get_states(self):
-        self.ratio_hold = self.num_stocks / \
-            int(self.portfolio_value / self.environment.get_price())
-        self.ratio_portfolio_value = (
-            self.portfolio_value / self.base_portfolio_value
-        )
+        self.ratio_hold = self.num_stocks * self.environment.get_price() \
+            / self.portfolio_value
         return (
             self.ratio_hold,
-            self.ratio_portfolio_value,
-            (self.environment.get_price() / self.avg_buy_price) -
-            1 if self.avg_buy_price > 0 else 0
+            self.profitloss,
+            (self.environment.get_price() / self.avg_buy_price) - 1 \
+                if self.avg_buy_price > 0 else 0
         )
 
     def decide_action(self, pred_value, pred_policy, epsilon):
@@ -104,10 +91,7 @@ class Agent:
         # 탐험 결정
         if np.random.rand() < epsilon:
             exploration = True
-            if np.random.rand() < self.exploration_base:
-                action = self.ACTION_BUY
-            else:
-                action = np.random.randint(self.NUM_ACTIONS - 1) + 1
+            action = np.random.randint(self.NUM_ACTIONS)
         else:
             exploration = False
             action = np.argmax(pred)
@@ -135,12 +119,10 @@ class Agent:
     def decide_trading_unit(self, confidence):
         if np.isnan(confidence):
             return self.min_trading_unit
-        added_traiding = max(min(
-            int(confidence * (self.max_trading_unit -
-                self.min_trading_unit)),
-            self.max_trading_unit-self.min_trading_unit
-        ), 0)
-        return self.min_trading_unit + added_traiding
+        added_trading = max(min(
+            int(confidence * (self.max_trading_unit - self.min_trading_unit)),
+            self.max_trading_unit-self.min_trading_unit), 0)
+        return self.min_trading_unit + added_trading
 
     def act(self, action, confidence):
         if not self.validate_action(action):
@@ -162,20 +144,17 @@ class Agent:
             )
             # 보유 현금이 모자랄 경우 보유 현금으로 가능한 만큼 최대한 매수
             if balance < 0:
-                trading_unit = max(
-                    min(
-                        int(self.balance / (
-                            curr_price * (1 + self.TRADING_CHARGE))),
-                        self.max_trading_unit
-                    ),
-                    self.min_trading_unit
-                )
+                trading_unit = max(min(
+                    int(self.balance / (curr_price * (1 + self.TRADING_CHARGE))),
+                    self.max_trading_unit
+                ), self.min_trading_unit)
             # 수수료를 적용하여 총 매수 금액 산정
             invest_amount = curr_price * \
                 (1 + self.TRADING_CHARGE) * trading_unit
             if invest_amount > 0:
-                self.avg_buy_price = (self.avg_buy_price * self.num_stocks +
-                                      curr_price) / (self.num_stocks + trading_unit)  # 주당 매수 단가 갱신
+                self.avg_buy_price = \
+                    (self.avg_buy_price * self.num_stocks + curr_price) \
+                        / (self.num_stocks + trading_unit)  # 주당 매수 단가 갱신
                 self.balance -= invest_amount  # 보유 현금을 갱신
                 self.num_stocks += trading_unit  # 보유 주식 수를 갱신
                 self.num_buy += 1  # 매수 횟수 증가
@@ -190,8 +169,11 @@ class Agent:
             invest_amount = curr_price * (
                 1 - (self.TRADING_TAX + self.TRADING_CHARGE)) * trading_unit
             if invest_amount > 0:
-                self.avg_buy_price = (self.avg_buy_price * self.num_stocks - curr_price) / (
-                    self.num_stocks - trading_unit) if self.num_stocks > trading_unit else 0  # 주당 매수 단가 갱신
+                # 주당 매수 단가 갱신
+                self.avg_buy_price = \
+                    (self.avg_buy_price * self.num_stocks - curr_price) \
+                        / (self.num_stocks - trading_unit) \
+                            if self.num_stocks > trading_unit else 0
                 self.num_stocks -= trading_unit  # 보유 주식 수를 갱신
                 self.balance += invest_amount  # 보유 현금을 갱신
                 self.num_sell += 1  # 매도 횟수 증가
@@ -202,8 +184,5 @@ class Agent:
 
         # 포트폴리오 가치 갱신
         self.portfolio_value = self.balance + curr_price * self.num_stocks
-        self.profitloss = (
-            (self.portfolio_value - self.initial_balance) / self.initial_balance
-        )
-
+        self.profitloss = self.portfolio_value / self.initial_balance - 1
         return self.profitloss
