+++ quantylab/rltrader/environment.py
@@ -21,6 +21,3 @@ class Environment:
         if self.observation is not None:
             return self.observation[self.PRICE_IDX]
         return None
-
-    def set_chart_data(self, chart_data):
-        self.chart_data = chart_data
