+++ quantylab/rltrader/networks_pytorch.py
@@ -1,0 +1,215 @@
+import threading
+import numpy as np
+
+import torch
+
+
+class Network:
+    lock = threading.Lock()
+
+    def __init__(self, input_dim=0, output_dim=0, lr=0.001, 
+                shared_network=None, activation='sigmoid', loss='mse'):
+        self.input_dim = input_dim
+        self.output_dim = output_dim
+        self.lr = lr
+        self.shared_network = shared_network
+        self.activation = activation
+        self.loss = loss
+        
+        inp = (self.input_dim,)
+        self.head = None
+        if self.shared_network is None:
+            self.head = self.get_network_head(inp, self.output_dim)
+        else:
+            self.head = self.shared_network
+
+        self.model = torch.nn.Sequential(self.head)
+        if self.activation == 'linear':
+            pass
+        elif self.activation == 'relu':
+            self.model.add_module('activation', torch.nn.ReLU())
+        elif self.activation == 'sigmoid':
+            self.model.add_module('activation', torch.nn.Sigmoid())
+        elif self.activation == 'tanh':
+            self.model.add_module('activation', torch.nn.Tanh())
+        self.model.apply(Network.init_weights)
+
+        self.optimizer = torch.optim.Adam(self.model.parameters(), lr=self.lr)
+        self.criterion = None
+        if loss == 'mse':
+            self.criterion = torch.nn.MSELoss()
+        elif loss == 'binary_crossentropy':
+            self.criterion = torch.nn.BCELoss()
+
+    def predict(self, sample):
+        with self.lock:
+            self.model.eval()
+            with torch.no_grad():
+                pred = self.model(torch.from_numpy(sample).float()).detach().numpy()
+                pred = pred.flatten()
+            return pred
+
+    def train_on_batch(self, x, y):
+        self.model.train()
+        loss = 0.
+        with self.lock:
+            y_pred = self.model(torch.from_numpy(x).float())
+            _loss = self.criterion(y_pred, torch.from_numpy(y).float())
+            self.optimizer.zero_grad()
+            _loss.backward()
+            self.optimizer.step()
+            loss += _loss.item()
+        return loss
+
+    def save_model(self, model_path):
+        if model_path is not None and self.model is not None:
+            torch.save(self.model, model_path)
+
+    def load_model(self, model_path):
+        if model_path is not None:
+            self.model = torch.load(model_path)
+
+    @classmethod
+    def get_shared_network(cls, net='dnn', num_steps=1, input_dim=0, output_dim=0):
+        if net == 'dnn':
+            return DNN.get_network_head((input_dim,), output_dim)
+        elif net == 'lstm':
+            return LSTMNetwork.get_network_head((num_steps, input_dim), output_dim)
+        elif net == 'cnn':
+            return CNN.get_network_head((num_steps, input_dim), output_dim)
+        elif net == 'q3':
+            return Q3.get_network_head((num_steps, input_dim), output_dim)
+
+    # @staticmethod
+    def init_weights(m):
+        if isinstance(m, torch.nn.Linear) or isinstance(m, torch.nn.Conv1d):
+            torch.nn.init.normal_(m.weight, std=0.05)
+        elif isinstance(m, torch.nn.LSTM):
+            for weights in m.all_weights:
+                for weight in weights:
+                    torch.nn.init.normal_(weight, std=0.05)
+
+class DNN(Network):
+    @staticmethod
+    def get_network_head(inp, output_dim):
+        return torch.nn.Sequential(
+            torch.nn.BatchNorm1d(inp[0]),
+            torch.nn.Linear(inp[0], 256),
+            torch.nn.Dropout(p=.1),
+            torch.nn.Linear(256, 128),
+            torch.nn.Dropout(p=.1),
+            torch.nn.Linear(128, 64),
+            torch.nn.Dropout(p=.1),
+            torch.nn.Linear(64, 32),
+            torch.nn.Dropout(p=.1),
+            torch.nn.Linear(32, output_dim),
+        )
+
+    def train_on_batch(self, x, y):
+        x = np.array(x).reshape((-1, self.input_dim))
+        return super().train_on_batch(x, y)
+
+    def predict(self, sample):
+        sample = np.array(sample).reshape((1, self.input_dim))
+        return super().predict(sample)
+
+
+class LSTMNetwork(Network):
+    def __init__(self, *args, num_steps=1, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.num_steps = num_steps
+
+    @staticmethod
+    def get_network_head(inp, output_dim):
+        return torch.nn.Sequential(
+            torch.nn.BatchNorm1d(inp[0]),
+            LSTMModule(inp[1], 32, batch_first=True, use_last_only=True),
+            torch.nn.BatchNorm1d(32),
+            torch.nn.Dropout(p=.1),
+            torch.nn.Linear(32, output_dim),
+        )
+
+    def train_on_batch(self, x, y):
+        x = np.array(x).reshape((-1, self.num_steps, self.input_dim))
+        return super().train_on_batch(x, y)
+
+    def predict(self, sample):
+        sample = np.array(sample).reshape((-1, self.num_steps, self.input_dim))
+        return super().predict(sample)
+
+
+class CNN(Network):
+    def __init__(self, *args, num_steps=1, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.num_steps = num_steps
+
+    @staticmethod
+    def get_network_head(inp, output_dim):
+        kernel_size = 2
+        return torch.nn.Sequential(
+            torch.nn.BatchNorm1d(inp[0]),
+            torch.nn.Conv1d(inp[0], 5, kernel_size),
+            torch.nn.BatchNorm1d(5),
+            torch.nn.Dropout(p=.1),
+            torch.nn.Conv1d(5, 1, kernel_size),
+            torch.nn.BatchNorm1d(1),
+            torch.nn.Flatten(),
+            torch.nn.Dropout(p=.1),
+            torch.nn.Linear(inp[1] - (kernel_size - 1) * 2, output_dim),
+        )
+
+    def train_on_batch(self, x, y):
+        x = np.array(x).reshape((-1, self.num_steps, self.input_dim))
+        return super().train_on_batch(x, y)
+
+    def predict(self, sample):
+        sample = np.array(sample).reshape((1, self.num_steps, self.input_dim))
+        return super().predict(sample)
+
+
+class Q3(Network):
+    def __init__(self, *args, num_steps=1, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.num_steps = num_steps
+
+    @staticmethod
+    def get_network_head(inp, output_dim):
+        kernel_size = 2
+        return torch.nn.Sequential(
+            torch.nn.BatchNorm1d(inp[0]),
+            torch.nn.Conv1d(inp[0], 5, kernel_size),
+            torch.nn.BatchNorm1d(5),
+            LSTMModule(inp[1] - kernel_size + 1, 256, batch_first=True),
+            torch.nn.BatchNorm1d(256),
+            torch.nn.Dropout(p=.1),
+            torch.nn.Linear(256, 128),
+            torch.nn.BatchNorm1d(128),
+            torch.nn.Dropout(p=.1),
+            torch.nn.Linear(128, 64),
+            torch.nn.BatchNorm1d(64),
+            torch.nn.Dropout(p=.1),
+            torch.nn.Linear(64, 32),
+            torch.nn.BatchNorm1d(32),
+            torch.nn.Dropout(p=.1),
+            torch.nn.Linear(32, output_dim),
+        )
+
+    def train_on_batch(self, x, y):
+        x = np.array(x).reshape((-1, self.num_steps, self.input_dim))
+        return super().train_on_batch(x, y)
+
+    def predict(self, sample):
+        sample = np.array(sample).reshape((-1, self.num_steps, self.input_dim))
+        return super().predict(sample)
+
+
+class LSTMModule(torch.nn.LSTM):
+    def __init__(self, *args, use_last_only=False, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.use_last_only = use_last_only
+
+    def forward(self, x):
+        output, (h_n, _) = super().forward(x)
+        if self.use_last_only:
+            return h_n[-1]
+        return output
